apps/mcp_platform/requirements.txt
# MCP Platform - Unified Requirements
# This file consolidates requirements from all MCP components

# Core Kailash SDK
kailash-sdk>=0.6.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
jinja2>=3.1.0

# MCP Protocol Support - Official Anthropic Package Only
mcp>=1.0.0

# Async and Networking
aiohttp>=3.9.0
websockets>=11.0
httpx>=0.25.0
tenacity>=8.2.0
circuit-breaker>=1.4.0

# Database and ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
alembic>=1.12.0

# Caching and Message Queue
redis>=5.0.0
celery>=5.3.0
kombu>=5.3.0

# Security and Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
pyotp>=2.8.0
cryptography>=41.0.0
authlib>=1.2.0

# Rate Limiting
slowapi>=0.1.9
limits>=3.6.0

# Monitoring and Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-exporter-jaeger>=1.21.0
sentry-sdk>=1.38.0
structlog>=23.1.0

# Configuration and CLI
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0

# Utilities
psutil>=5.9.0
orjson>=3.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
factory-boy>=3.3.0

# Development Tools
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0
isort>=5.12.0
pre-commit>=3.4.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0
mkdocstrings[python]>=0.24.0
